+++ atomina-test/AtominaTestGame.cpp
@@ -11,7 +11,7 @@ void GameTest::setup(ATMA::ATMAContext *l_ctx)
     active = true;
 }
 
-void GameTest::update(ATMA::ATMAContext *l_ctx, const long long &l_dt)
+void GameTest::update(ATMA::ATMAContext *l_ctx, const double &l_dt)
 {
     ATMA_ENGINE_INFO("Test Wrapper is now running");
     ATMA_ENGINE_INFO("Running Test Game with argc: {}", *l_ctx->argc);
